+++ src/main/store/store.ts
@@ -7,6 +7,7 @@ import { migrateSettingsV5toV6 } from 'src/main/store/migrations/v5-to-v6';
 
 import logger from '../logger';
 
+import { migrateV6ToV7 } from './migrations/v6-to-v7';
 import { migrateSettingsV0toV1 } from './migrations/v0-to-v1';
 import { migrateSettingsV1toV2 } from './migrations/v1-to-v2';
 import { migrateSettingsV2toV3 } from './migrations/v2-to-v3';
@@ -111,7 +112,7 @@ interface StoreSchema {
   userId?: string;
 }
 
-const CURRENT_SETTINGS_VERSION = 6;
+const CURRENT_SETTINGS_VERSION = 7;
 
 interface CustomStore<T> {
   get<K extends keyof T>(key: K): T[K] | undefined;
@@ -226,6 +227,11 @@ export class Store {
       if (settingsVersion === 5) {
         settings = migrateSettingsV5toV6(settings);
         settingsVersion = 6;
+      }
+
+      if (settingsVersion === 6) {
+        settings = migrateV6ToV7(settings);
+        settingsVersion = 7;
       }
 
       this.store.set('settings', settings as SettingsData);